- changed maxRaritySum from double to int - made sure that when there are maxRaritySum different rare features - they get positive score (before change - maxRaritySum-1 different rare features got positive score)

+++ fortscale/fortscale-ml/src/main/java/fortscale/ml/model/builder/CategoryRarityModelBuilder.java
@@ -18,12 +18,12 @@ public class CategoryRarityModelBuilder implements IModelBuilder {
 
     private Pattern ignoreValues;
     private int maxPossibleRarity;
-    private double maxRaritySum;
+    private int maxRaritySum;
 
     @JsonCreator
     public CategoryRarityModelBuilder(@JsonProperty("ignorePattern") String ignorePattern,
                                       @JsonProperty("maxPossibleRarity") Integer maxPossibleRarity,
-                                      @JsonProperty("maxRaritySum") Double maxRaritySum) {
+                                      @JsonProperty("maxRaritySum") Integer maxRaritySum) {
         Assert.notNull(maxPossibleRarity);
         Assert.isTrue(maxPossibleRarity > 0);
         Assert.notNull(maxRaritySum);
