+++ fortscale/fortscale-ml/src/main/java/fortscale/ml/model/prevalance/field/CategoryRarityModel.java
@@ -13,7 +13,7 @@ public class CategoryRarityModel implements Model {
 
 	private RarityScorer occurrencesHistogram;
 
-	public CategoryRarityModel(Collection<Integer> featureCounts, int maxPossibleRarity, double maxRaritySum) {
+	public CategoryRarityModel(Collection<Integer> featureCounts, int maxPossibleRarity, int maxRaritySum) {
 		occurrencesHistogram = new RarityScorer(featureCounts, maxPossibleRarity, maxRaritySum);
 	}
